apm-1017: check PR numbers availability
