+++ docker/async-slowapp/.mocharc.yml
@@ -1,0 +1,4 @@
+spec: './src/*.spec.js'
+ui: bdd
+watch-files:
+  - 'src/**/*.js'
