+++ docker/async-slowapp/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "sync-wrap",
+  "name": "async-slowapp",
   "version": "1.0.0",
   "main": "src/main.js",
   "scripts": {
@@ -9,9 +9,9 @@
   "author": "",
   "license": "",
   "dependencies": {
-    "agentkeepalive": "^4.1.2",
     "express": "^4.17.1",
-    "loglevel": "^1.6.8"
+    "loglevel": "^1.6.8",
+    "uuid": "^8.0.0"
   },
   "devDependencies": {
     "chai": "^4 .2.0",
