--- docker/sync-wrap-async-slowapp/src/main.js
@@ -1,11 +1,0 @@
-"use strict";
-
-const app = require("./app");
-
-
-app.setup(process.env);
-
-
-const server = app.start(process.env);
-
-module.exports = {server: server};
