--- ecs-proxies.yml
@@ -1,10 +1,0 @@
-
-docker_containers:
-  - name: sync-wrap
-    dockerfile: "docker/sync-wrap/Dockerfile"
-    path: "docker/sync-wrap"
-    port: 9000
-  - name: sync-wrap-async-slowapp
-    dockerfile: "docker/sync-wrap-async-slowapp/Dockerfile"
-    path: "docker/sync-wrap-async-slowapp"
-    port: 9000
